+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @m5kdev/backend
 
+## 0.2.0
+
+### Minor Changes
+
+- trcp rework
+
+### Patch Changes
+
+- Updated dependencies
+  - @m5kdev/commons@0.2.0
+  - @m5kdev/config@0.2.0
+
+## 0.1.5
+
+### Patch Changes
+
+- declarations
+- Updated dependencies
+  - @m5kdev/commons@0.1.5
+  - @m5kdev/config@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- build added
+- Updated dependencies
+  - @m5kdev/commons@0.1.4
+  - @m5kdev/config@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes