+++ packages/email/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @m5kdev/email
 
+## 0.2.0
+
+### Minor Changes
+
+- trcp rework
+
+## 0.1.5
+
+### Patch Changes
+
+- declarations
+
+## 0.1.4
+
+### Patch Changes
+
+- build added
+
 ## 0.1.3
 
 ### Patch Changes